requirements.txt
streamlit
pandas
scikit-learn
pyttsx3
gtts
googletrans==4.0.0-rc1
Pillow

